Attachments search/clone; support for dbd-6.5.2; bump version
* update version to 0.3.0
* fix confirm default capitalization
* fix json encoding issue
* rename --launcher to --launch
* add menu item aliases
* convert command runner into class
* update README for searching and cloning attachments
* update settings file template
* add umodel integration for search/visualize/saving assets
* add keyboard listener for stopping long operations

+++ dbdmodswap/helpers/tempFileHelpers.py
@@ -4,16 +4,21 @@ from contextlib import contextmanager
 
 
 @contextmanager
-def openTemporaryFile(dir=None, prefix=None, suffix=None, mode=None):
+def openTemporaryFile(dir=None, prefix=None, suffix=None, mode=None, encoding=None, deleteFirst=False):
     file = tempfile.NamedTemporaryFile(
+        mode or 'w',
         dir=dir,
         prefix=prefix,
         suffix=suffix,
-        mode=mode,
+        encoding=encoding or 'utf-8',
         delete=False,
     )
     try:
+        if deleteFirst:
+            file.close()
+            pathlib.Path.unlink(file.name)
+
         yield file
     finally:
         file.close()
-        pathlib.Path.unlink(file.name)
+        pathlib.Path.unlink(file.name, missing_ok=True)
